requirements.txt
appdirs==1.4.3
certifi==2017.1.23
future==0.16.0
gTTS==1.1.8
gTTS-token==1.1.1
packaging==16.8
pyparsing==2.2.0
python-telegram-bot==5.3.0
requests==2.13.0
six==1.10.0
urllib3==1.20
